+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.0.9000.9001
+* Renamed `refDoseBeta` slot to `ref_dose_beta` in `DualEndpointBeta` class.
 * Replaced `smooth` character vector argument of `DualEndpointRW()` to `rw1` flag. 
 * Added new `ModelParamsNormal` class to represent parameters of bivariate normal
   distribution.